Skip ruby init if not installed

+++ zshrc
@@ -32,9 +32,11 @@ stty -ixon
 [[ -e ~/.bash_aliases ]] && . ~/.bash_aliases
 [[ -e ~/.api_keys ]] && . ~/.api_keys
 
-# saves 100ms in total, rm ~/.gem/ruby/current when updating system ruby
-[[ /var/log/pacman.log -nt ~/.gem/ruby/current.touch ]] && rm ~/.gem/ruby/current
-[[ -e ~/.rbenv_init ]] && . ~/.rbenv_init
+if command -v ruby >/dev/null 2>&1; then
+  # saves 100ms in total, rm ~/.gem/ruby/current when updating system ruby
+  [[ /var/log/pacman.log -nt ~/.gem/ruby/current.touch ]] && rm ~/.gem/ruby/current
+  [[ -e ~/.rbenv_init ]] && . ~/.rbenv_init
+fi
 
 export PATH=$HOME/bin:$HOME/bin/games:$HOME/code/scripts:$HOME/.cargo/bin:$PATH
